requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
python-multipart
